requirements.txt
Flask>=2.3.0
pandas>=2.0.0
openpyxl>=3.1.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
requests>=2.31.0 